// Logitude/AngularModules/CypressTesting/Maintenance/cypress/models/SharedLogisticsDetails/PartnersPermissionsDetails.ts
export interface PartnersPermissionsDetails {

    Shipper: string
    Consignee: string
    Agent: string
    ShipperNotExporter: string
    ConsigneeNotImporter: string
    Notify1: string
    Notify2: string
    FreightForwarder: string
    Coloader: string
    CustomAgentExport: string
    CustomAgentImport: string
    CustomClearancePoint: string
    Consolidator: string
    ReleasingAgent: string
    IssuingCarrierAgent: string
}